Move TSDocConfiguration under src/configuration
