+++ tsdoc/src/nodes/DocNodeContainer.ts
@@ -15,8 +15,8 @@ export interface IDocNodeContainerParsedParameters extends IDocNodeParsedParamet
 }
 
 /**
- * DocNodeContainer is the base class for DocNode classes that act as a simple container
- * for other child nodes.  The child classes are {@link DocParagraph} and {@link DocSection}.
+ * DocNodeContainer is the base class for DocNode classes that allow arbitrary child nodes to be added by the consumer.
+ * The child classes are {@link DocParagraph} and {@link DocSection}.
  */
 export abstract class DocNodeContainer extends DocNode {
   private readonly _nodes: DocNode[] = [];
